add aws-apigv1-stream example, revert aws-lambda-stream changes
