+++ examples/aws-apigv1-stream/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "aws-apigv1-stream",
+  "version": "1.0.0",
+  "dependencies": {
+    "sst": "^4"
+  }
+}
