+++ examples/aws-lambda-stream/sst.config.ts
@@ -54,7 +54,7 @@
  * curl --no-buffer https://u3dyblk457ghskwbmzrbylpxoi0ayrbb.lambda-url.us-east-1.on.aws
  * ```
  *
- * Streaming is also supported through API Gateway REST API.
+ * Here we are using a Function URL directly because API Gateway doesn't support streaming.
  *
  */
 export default $config({
@@ -73,15 +73,8 @@ export default $config({
       handler: "index.handler",
     });
 
-    const api = new sst.aws.ApiGatewayV1("MyApi");
-    api.route("GET /", "index.handler", {
-      streaming: true,
-    });
-    api.deploy();
-
     return {
       url: fn.url,
-      api: api.url,
     };
   },
 });
